Bump nixplgs and GHC to 9.8.2
One unfortunate fallout from this bump was that cabal 3.12 broke the use
of `active-repositories: none` which was used to tell Cabal not to
manage package repositories (as we use Nix for that).

The fix for now is to patch `cabal-install` and revert the change that
made the previoud warning into an error.